backend/requirements.txt
flask>=2.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
flask-cors>=4.0.0
